개별 테스트 작업 완료

+++ src/main/kotlin/com/blife/blife/infra/external/booksearchapi/BookSearchService.kt
@@ -25,7 +25,7 @@ class BookSearchService(
 		return null
 	}
 
-	fun searchBookListByTitle(title: String, page: Int): List<Book>? {
+	fun searchBookListByTitle(title: String, page: Long): List<Book>? {
 		for (client in bookSearchClients) {
 			client.searchBookListByTitle(title, page).let {
 				if (it.first != null) return it.first
